Fixed DI and settings
