+++ settings.yaml
@@ -47,3 +47,13 @@ SmsNotifications:
                 - AzureTableStorage
         QueueName:
             settings-key: SmsNotifications-QueueName
+ClientAccountServiceClient: 
+    ServiceUrl:
+        settings-key: ClientAccountServiceClient_ServiceUrl
+PersonalDataServiceClient:
+    ServiceUrl:
+        settings-key: PersonalDataServiceClient_ServiceUrl
+    ServiceExternalUri:
+        settings-key: PersonalDataServiceClient_ServiceExternalUri
+    ApiKey:
+        settings-key: PersonalDataServiceClient_ApiKey
