+++ src/Lykke.Service.KycNotifications/Settings/AppSettings.cs
@@ -1,6 +1,8 @@
 ﻿using JetBrains.Annotations;
 using Lykke.Sdk.Settings;
+using Lykke.Service.ClientAccount.Client;
 using Lykke.Service.KycNotifications.Core.Settings;
+using Lykke.Service.PersonalData.Settings;
 
 namespace Lykke.Service.KycNotifications.Settings
 {
@@ -11,5 +13,7 @@ namespace Lykke.Service.KycNotifications.Settings
 		public LykkeKycWebsiteUrlSettings LykkeKycWebsiteUrlSettings { get; set; }
 		public SmsNotificationsSettings SmsNotifications { get; set; }
         public PushNotificationsServiceSettings PushNotificationsService { get; set; }
+		public PersonalDataServiceClientSettings PersonalDataServiceClient { get; set; }
+        public ClientAccountServiceClientSettings ClientAccountServiceClient { get; set; }
     }
 }
